python/packages/corlinman-providers/pyproject.toml
[project]
name = "corlinman-providers"
version = "0.1.0"
description = "LLM provider adapters (Anthropic / OpenAI / Google / China) + failover classification."
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "corlinman contributors" }]

dependencies = [
    "anthropic>=0.39",
    "openai>=1.52",
    "google-genai>=0.8",
    "httpx>=0.27",
    "pydantic>=2.10",
    "structlog>=25.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corlinman_providers"]
